fix(asynciterable-pipe): fix asynciterable pipe typings
fix #265
