+++ src/Ix.dom.ts
@@ -16,72 +16,11 @@ export { OrderedAsyncIterable } from './Ix';
 export { OrderedAsyncIterableBase } from './Ix';
 export { AsyncSink, Iterable, AsyncIterable } from './Ix';
 
-import { OperatorAsyncFunction } from './interfaces';
-
-declare module './asynciterable/asynciterablex' {
-  interface AsyncIterableX<T> {
-    pipe(): AsyncIterableX<T>;
-    pipe<A>(op1: OperatorAsyncFunction<T, A>): AsyncIterableX<A>;
-    pipe<A, B>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>
-    ): AsyncIterableX<B>;
-    pipe<A, B, C>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>
-    ): AsyncIterableX<C>;
-    pipe<A, B, C, D>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>
-    ): AsyncIterableX<D>;
-    pipe<A, B, C, D, E>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>,
-      op5: OperatorAsyncFunction<D, E>
-    ): AsyncIterableX<E>;
-    pipe<A, B, C, D, E, F>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>,
-      op5: OperatorAsyncFunction<D, E>,
-      op6: OperatorAsyncFunction<E, F>
-    ): AsyncIterableX<F>;
-    pipe<A, B, C, D, E, F, G>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>,
-      op5: OperatorAsyncFunction<D, E>,
-      op6: OperatorAsyncFunction<E, F>,
-      op7: OperatorAsyncFunction<F, G>
-    ): AsyncIterableX<G>;
-    pipe<A, B, C, D, E, F, G, H>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>,
-      op5: OperatorAsyncFunction<D, E>,
-      op6: OperatorAsyncFunction<E, F>,
-      op7: OperatorAsyncFunction<F, G>,
-      op8: OperatorAsyncFunction<G, H>
-    ): AsyncIterableX<H>;
-    pipe<A, B, C, D, E, F, G, H, I>(
-      op1: OperatorAsyncFunction<T, A>,
-      op2: OperatorAsyncFunction<A, B>,
-      op3: OperatorAsyncFunction<B, C>,
-      op4: OperatorAsyncFunction<C, D>,
-      op5: OperatorAsyncFunction<D, E>,
-      op6: OperatorAsyncFunction<E, F>,
-      op7: OperatorAsyncFunction<F, G>,
-      op8: OperatorAsyncFunction<G, H>,
-      op9: OperatorAsyncFunction<H, I>
-    ): AsyncIterableX<I>;
-    pipe<R>(...operations: OperatorAsyncFunction<T, R>[]): AsyncIterableX<R>;
-  }
-}
+import './add/asynciterable/fromdomstream';
+import './add/asynciterable-operators/todomstream';
+export { toDOMStream } from './asynciterable/todomstream';
+export {
+  fromDOMStream,
+  AsyncIterableReadableStream,
+  AsyncIterableReadableByteStream
+} from './asynciterable/fromdomstream';
